Change tx processing
Add tag field to Invoice and Payment
Add reference to BPSManager to InvoiceProcessor and PaymentProcessor
Implement getTx methods for CoinClients and BPSManager

+++ src/main/kotlin/dsx/bps/crypto/xrp/datamodel/XrpAccountData.kt
@@ -8,5 +8,5 @@ data class XrpAccountData(
     @SerializedName("Balance")
     val balance: String,
     @SerializedName("Sequence")
-    val sequence: Long
+    val sequence: Int
 )
